feat (test microservice): added endpoints to retrieve general and specific tests

+++ test-service/src/main/java/com/unicauca/sga/testService/Infrastructure/Controllers/ManageQuestionsController/DTOs/Request/QuestionDTORequest.java
@@ -27,5 +27,5 @@ public class QuestionDTORequest {
     private String questionStructure;
 
     @NotNull(message = "Debe especificar la evaluación a la que pertenece la pregunta")
-    private Long testId;
+    private Integer testId;
 }
